Move model & inference out of /api and back to /src for local development
