+++ .gitignore
@@ -10,4 +10,5 @@ ios/wavescore.xcodeproj/
 ios/wavescore/.DS_Store
 models/
 src/__pycache__/
+src/keys/
 tmp/
